refactored match history into helper
