+++ client/league.py
@@ -3,11 +3,7 @@ from helper.league import LeagueClientHelper
 from utils.logger import Logger
 import re
 from cassiopeia import Summoner
-from cassiopeia.data import GameMode
 
-def format_game_mode(mode : GameMode) -> str:
-    # "GameMode.aram" -> "ARAM"
-    return str(mode)[9:].upper()
 class LeagueClient:
     @staticmethod
     async def display_game_summary(message : Message) -> None:
@@ -43,7 +39,6 @@ class LeagueClient:
     async def display_match_history(message : Message) -> None:
         summoner_name = message.content[14:]
         summoner = Summoner(name=summoner_name, region="NA")
-        history = summoner.match_history[:5]
         output = "{:>8} {:>14} {:>14} {:>14} {:>15} {:>16}".format(
                                                     "MODE",
                                                     "CHAMPION",
@@ -51,20 +46,8 @@ class LeagueClient:
                                                     "DEATHS",
                                                     "ASSISTS",
                                                     "KDA") + "\n" 
-        for match in history:
-            for player in match.participants:
-                if player.summoner.id == summoner.id:
-                    kills = player.stats.kills
-                    deaths = player.stats.deaths
-                    assists = player.stats.assists
-                    kda = str(round(player.stats.kda, 2))
-                    output += "{:>8} {:>14} {:>14} {:>14} {:>15} {:>16}".format(
-                                                                format_game_mode(match.mode),
-                                                                player.champion.name,
-                                                                kills,
-                                                                deaths,
-                                                                assists,
-                                                                kda) + "\n" 
+        
+        output +=  LeagueClientHelper.display_match_history_info(summoner, summoner.match_history[:5])
                     
         await message.channel.send(f'**Last 5 Games for {summoner.name}:**```{output}```')
         Logger.log("Outputting Match History")
